+++ src/address.h
@@ -49,7 +49,7 @@ class IPAddress : public Address {
 public:
     typedef std::shared_ptr<IPAddress> ptr;
 
-    static IPAddress::ptr Create(const char* address, uint32_t port = 0);
+    static IPAddress::ptr Create(const char* address, uint16_t port = 0);
 
     virtual IPAddress::ptr broadcastAddress(uint32_t prefix_len) = 0;
     virtual IPAddress::ptr networdAddress(uint32_t prefix_len) = 0;
@@ -63,10 +63,10 @@ class IPv4Address : public IPAddress {
 public:
     typedef std::shared_ptr<IPv4Address> ptr;
 
-    static IPv4Address::ptr Create(const char* address, uint32_t port = 0);
+    static IPv4Address::ptr Create(const char* address, uint16_t port = 0);
 
     IPv4Address(const struct sockaddr_in& address);
-    IPv4Address(uint32_t address = INADDR_ANY, uint32_t port = 0); // INADDR_ANY 转换 0.0.0.0
+    IPv4Address(uint32_t address = INADDR_ANY, uint16_t port = 0); // INADDR_ANY 转换 0.0.0.0
 
     const sockaddr* getAddr() const override;
     socklen_t getAddrLen() const override;
@@ -87,11 +87,11 @@ class IPv6Address : public IPAddress {
 public:
     typedef std::shared_ptr<IPv6Address> ptr;
 
-    static IPv6Address::ptr Create(const char* address, uint32_t port = 0);
+    static IPv6Address::ptr Create(const char* address, uint16_t port = 0);
 
     IPv6Address();
     IPv6Address(const struct sockaddr_in6& address);
-    IPv6Address(const uint8_t address[16], uint32_t port = 0); // INADDR_ANY 转换 0.0.0.0
+    IPv6Address(const uint8_t address[16], uint16_t port = 0); // INADDR_ANY 转换 0.0.0.0
 
     const sockaddr* getAddr() const override;
     socklen_t getAddrLen() const override;
